# Use BroadcastKernel and ReduceKernel to optimize expand and expand_grad.